refactor: add mahalanobis and padim modules
